refactor(ui): #173 move Box to @codeimage/ui
