+++ apps/codeimage/src/components/Footer/Footer.tsx
@@ -1,7 +1,6 @@
 import * as styles from './Footer.css';
-import {Box} from '../../ui/Box/Box';
+import {Box, Link} from '@codeimage/ui';
 import {appEnvironment} from '../../core/configuration';
-import {Link} from '@codeimage/ui';
 
 export const Footer = () => {
   const {version} = appEnvironment;
