+++ apps/codeimage/src/components/Frame/Frame.tsx
@@ -3,7 +3,7 @@ import * as styles from './Frame.css';
 import {assignInlineVars} from '@vanilla-extract/dynamic';
 import {createHorizontalResize} from '../../core/hooks/resizable';
 import {exportExclude as _exportExclude} from '../../core/directives/exportExclude';
-import {Box} from '../../ui/Box/Box';
+import {Box} from '@codeimage/ui';
 import {FadeInOutTransition} from '../../ui/Transition/Transition';
 import {useModality} from '../../core/hooks/isMobile';
