+++ apps/codeimage/src/components/PromptUpdate/PromptUpdate.tsx
@@ -1,8 +1,7 @@
 import {Component, createEffect} from 'solid-js';
 import {useRegisterSW} from 'virtual:pwa-register/solid';
 import {notificationStore} from '../../ui/Toast/SnackbarHost';
-import {Button} from '@codeimage/ui';
-import {Box} from '../../ui/Box/Box';
+import {Box, Button} from '@codeimage/ui';
 import {useI18n} from '@codeimage/locale';
 import {AppLocaleEntries} from '../../i18n';
