+++ apps/codeimage/src/components/PropertyEditor/EditorForm.tsx
@@ -1,6 +1,6 @@
 import * as styles from './EditorSidebar.css';
 import {Component} from 'solid-js';
-import {Box} from '../../ui/Box/Box';
+import {Box} from '@codeimage/ui';
 
 export const EditorForm: Component = props => (
   <Box class={styles.sidebar}>{props.children}</Box>
