+++ apps/codeimage/src/components/Scaffold/Sidebar/Sidebar.tsx
@@ -1,7 +1,7 @@
 import {Component} from 'solid-js';
 import * as styles from './Sidebar.css';
 import {SidebarVariants} from './Sidebar.css';
-import {Box} from '../../../ui/Box/Box';
+import {Box} from '@codeimage/ui';
 
 export const Sidebar: Component<SidebarVariants> = props => {
   return (
