+++ apps/codeimage/src/components/Terminal/TabIcon.tsx
@@ -1,6 +1,6 @@
 import {JSXElement, mergeProps} from 'solid-js';
 import {SvgExternalIcon} from '../../ui/SvgIcon/SvgExternalIcon';
-import {Box} from '../../ui/Box/Box';
+import {Box} from '@codeimage/ui';
 import * as styles from './terminal.css';
 import {LanguageIconDefinition} from '@codeimage/config';
