+++ apps/codeimage/src/components/TerminalControlField/TerminalControlField.tsx
@@ -3,7 +3,7 @@ import {Dynamic} from 'solid-js/web';
 import {terminal$} from '@codeimage/store/terminal';
 import {Group} from '../../ui/Group/Group';
 import {RadioBlock} from '../../ui/RadioBlock/RadioBlock';
-import {Box} from '../../ui/Box/Box';
+import {Box} from '@codeimage/ui';
 import {appEnvironment} from '../../core/configuration';
 import {fromObservableObject} from '../../core/hooks/from-observable-object';
