+++ apps/codeimage/src/components/Toolbar/ShareButton.tsx
@@ -2,7 +2,7 @@ import {Component, mergeProps, Show} from 'solid-js';
 import {Button} from '@codeimage/ui';
 import {useI18n} from '@codeimage/locale';
 import {AppLocaleEntries} from '../../i18n';
-import {Box} from '../../ui/Box/Box';
+import {Box} from '@codeimage/ui';
 import {SvgIcon} from '../../ui/SvgIcon/SvgIcon';
 import {appEnvironment} from '../../core/configuration';
 import {useHotkey} from '../../hooks/use-hotkey';
