+++ apps/codeimage/src/ui/Dialog/DialogPanel.tsx
@@ -8,7 +8,7 @@ import clsx from 'clsx';
 import * as styles from './Dialog.css';
 import {DialogPanelVariants} from './Dialog.css';
 import {PropsWithChildren} from 'solid-js/types/render/component';
-import {Box} from '../Box/Box';
+import {Box} from '@codeimage/ui';
 import {assignInlineVars} from '@vanilla-extract/dynamic';
 import {dynamicFullHeight} from '../../theme/base.css';
