+++ apps/codeimage/src/ui/Dropdown/DropdownPanel.tsx
@@ -2,7 +2,7 @@ import {Menu, MenuItem, PopoverPanel, PopoverPanelProps} from 'solid-headless';
 import {Component, Show} from 'solid-js';
 import * as styles from './Dropdown.css';
 import {Text} from '@codeimage/ui';
-import {Box} from '../Box/Box';
+import {Box} from '@codeimage/ui';
 
 type DropdownMenuPanel = PopoverPanelProps<'div'> & {
   title?: string;
