+++ apps/codeimage/src/ui/Group/Group.tsx
@@ -1,5 +1,5 @@
 import {Component} from 'solid-js';
-import {Box} from '../Box/Box';
+import {Box} from '@codeimage/ui';
 import * as styles from './Group.css';
 import {GroupVariants} from './Group.css';
