+++ packages/ui/src/index.ts
@@ -1,2 +1,9 @@
-export {colors, spacing, themeColors, themeVars, root} from './lib/theme';
+export {
+  colors,
+  spacing,
+  themeColors,
+  themeVars,
+  root,
+  sprinkles,
+} from './lib/theme';
 export * from './lib/primitives';
