+++ packages/ui/src/lib/theme/index.ts
@@ -2,3 +2,4 @@ export * from './colors';
 export * from './spacing';
 export * from './theme.css';
 export * from './global.css';
+export * from './sprinkles.css';
